Change PerformanceQueries to forward to PerformanceQueriesGeneral
